rename test directory to avoid conflicts
