+++ test/integration/notebook/test_status_graph_notebook.py
@@ -3,7 +3,7 @@ Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 SPDX-License-Identifier: Apache-2.0
 """
 
-from test.integration.graph_notebook.GraphNotebookIntegrationTest import GraphNotebookIntegrationTest
+from test.integration.notebook.GraphNotebookIntegrationTest import GraphNotebookIntegrationTest
 
 
 class TestGraphMagicStatus(GraphNotebookIntegrationTest):
